vllm_llama_inference_framework/cloud/draft_verifier.py
"""
F2: 云端 Draft 验证器
使用大模型验证并修正边端生成的 Draft tokens
"""
import asyncio
import time
from typing import List, Tuple, Dict, Any, Optional
import numpy as np

from common.types import (
    VerifyRequest, 
    VerifyResponse,
    DraftResponse
)


class DraftVerifier:
    """
    Draft 验证器 (基于 vLLM)
    
    实现 speculative decoding 的验证阶段:
    1. 接收边端的 Draft tokens
    2. 使用大模型并行验证
    3. 修正不合适的 tokens
    4. 返回验证后的结果
    """
    
    def __init__(
        self, 
        model_path: str,
        acceptance_threshold: float = 0.8
    ):
        self.model_path = model_path
        self.acceptance_threshold = acceptance_threshold
        
        # 加载 vLLM 模型
        self.model = self._load_model(model_path)
        self.kv_cache = {}
    
    def _load_model(self, model_path: str):
        """
        加载 vLLM 模型
        
        实际实现中，这里应该导入 vLLM:
        from vllm import LLM, SamplingParams
        """
        print(f"[Cloud] 加载 vLLM 模型: {model_path}")
        return MockVLLMModel(model_path)
    
    async def verify_draft(
        self, 
        request: VerifyRequest
    ) -> VerifyResponse:
        """
        验证 Draft tokens
        
        Args:
            request: 验证请求
            
        Returns:
            验证响应
        """
        start_time = time.time()
        
        # 1. 准备验证输入
        full_prompt = request.prompt
        draft_tokens = request.draft_tokens
        
        print(f"[Cloud] 验证 Draft: prompt='{full_prompt[:50]}...', "
              f"draft_tokens={len(draft_tokens)}")
        
        # 2. 使用大模型验证
        verified_tokens, corrected_positions = await self._verify_tokens_async(
            full_prompt,
            draft_tokens,
            request.max_verify_tokens
        )
        
        # 3. 计算接受率
        accepted_count = len(draft_tokens) - len(corrected_positions)
        total_count = len(draft_tokens)
        acceptance_rate = accepted_count / total_count if total_count > 0 else 0.0
        
        # 4. 生成最终文本
        final_text = request.prompt + ''.join(verified_tokens)
        
        # 5. 准备响应
        latency = (time.time() - start_time) * 1000
        
        return VerifyResponse(
            verified_tokens=verified_tokens,
            verified_token_ids=[],  # 实际使用时填充
            accepted_count=accepted_count,
            total_count=total_count,
            acceptance_rate=acceptance_rate,
            corrected_positions=corrected_positions,
            final_text=final_text,
            latency_ms=latency
        )
    
    async def _verify_tokens_async(
        self,
        prompt: str,
        draft_tokens: List[str],
        max_tokens: int
    ) -> Tuple[List[str], List[int]]:
        """
        异步验证 tokens
        
        实现 speculative decoding 的核心逻辑:
        1. 将 Draft tokens 作为候选
        2. 大模型并行计算每个位置的概率
        3. 根据概率决定是否接受 Draft token
        4. 在第一个拒绝的位置开始生成新 tokens
        """
        # 模拟异步验证
        await asyncio.sleep(0.02)  # 模拟延迟
        
        verified_tokens = []
        corrected_positions = []
        
        # 模拟验证过程
        for i, draft_token in enumerate(draft_tokens):
            # 模拟大模型对该位置的预测概率
            acceptance_prob = np.random.uniform(0.6, 0.98)
            
            if acceptance_prob > self.acceptance_threshold:
                # 接受 Draft token
                verified_tokens.append(draft_token)
            else:
                # 拒绝并修正
                corrected_positions.append(i)
                # 生成修正后的 token (这里简单模拟)
                corrected_token = f"[CORRECTED_{i}]"
                verified_tokens.append(corrected_token)
                
                # 从第一个拒绝的位置开始，后续都需要重新生成
                break
        
        # 如果需要继续生成更多 tokens
        if len(verified_tokens) < max_tokens:
            remaining_tokens = max_tokens - len(verified_tokens)
            for j in range(remaining_tokens):
                new_token = f"_cloud_token_{j}_"
                verified_tokens.append(new_token)
        
        return verified_tokens, corrected_positions
    
    def verify_with_speculative_decoding(
        self,
        prompt: str,
        draft_tokens: List[str],
        temperature: float = 0.0
    ) -> Tuple[List[str], List[int], Dict[str, Any]]:
        """
        使用 speculative decoding 算法验证
        
        这是更完整的实现，实际使用时需要接入 vLLM 的并行验证接口
        
        Args:
            prompt: 输入提示
            draft_tokens: Draft tokens
            temperature: 温度参数
            
        Returns:
            (verified_tokens, corrected_positions, debug_info)
        """
        # 这里应该使用 vLLM 的并行验证能力
        # 例如: vllm_model.verify_tokens(prompt, draft_tokens, temperature)
        
        debug_info = {
            'verification_method': 'speculative_decoding',
            'draft_length': len(draft_tokens),
            'temperature': temperature
        }
        
        # 模拟实现
        verified_tokens = []
        corrected_positions = []
        
        for i, draft_token in enumerate(draft_tokens):
            # 模拟验证逻辑
            if np.random.random() > 0.1:  # 90% 接受率
                verified_tokens.append(draft_token)
            else:
                # 拒绝并生成新 token
                corrected_positions.append(i)
                verified_tokens.append(f"[V_{i}]")
        
        return verified_tokens, corrected_positions, debug_info
    
    def batch_verify(
        self,
        requests: List[VerifyRequest]
    ) -> List[VerifyResponse]:
        """
        批量验证多个 Draft
        
        Args:
            requests: 验证请求列表
            
        Returns:
            验证响应列表
        """
        # 使用 vLLM 的批处理能力
        results = []
        
        for request in requests:
            # 这里应该使用 vLLM 的批量推理接口
            result = asyncio.run(self.verify_draft(request))
            results.append(result)
        
        return results
    
    def get_verification_stats(
        self,
        recent_requests: int = 100
    ) -> Dict[str, Any]:
        """获取验证统计"""
        # 这里应该维护一个请求历史队列
        # 模拟统计数据
        return {
            'avg_acceptance_rate': 0.85,
            'avg_latency_ms': 50.0,
            'total_requests': 1000,
            'recent_requests': recent_requests
        }


class MockVLLMModel:
    """模拟 vLLM 模型接口"""
    
    def __init__(self, model_path: str):
        self.model_path = model_path
        self.model_name = model_path.split('/')[-1]
        
    def generate(
        self,
        prompts: List[str],
        sampling_params: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """
        模拟 vLLM 批量生成接口
        
        实际 vLLM 返回格式:
        {
            'text': '生成的文本',
            'token_ids': [...],
            'logprobs': [...]
        }
        """
        results = []
        
        for prompt in prompts:
            # 模拟生成
            num_tokens = np.random.randint(10, 100)
            token_ids = np.random.randint(1000, 32000, num_tokens).tolist()
            text = ' '.join([f't_{tid}' for tid in token_ids])
            
            results.append({
                'text': text,
                'token_ids': token_ids,
                'logprobs': np.random.uniform(-2, -0.1, num_tokens).tolist()
            })
        
        return results
    
    def verify_tokens(
        self,
        prompt: str,
        draft_tokens: List[str],
        temperature: float = 0.0
    ) -> Dict[str, Any]:
        """
        模拟 vLLM 的 token 验证接口
        
        这是 speculative decoding 的关键接口
        """
        acceptance_probs = np.random.uniform(0.7, 0.99, len(draft_tokens))
        
        return {
            'acceptance_probs': acceptance_probs.tolist(),
            'verified_tokens': draft_tokens.copy(),
            'corrected_positions': []
        }


class VerificationEnsemble:
    """集成多个验证策略"""
    
    def __init__(self, verifiers: List[DraftVerifier] = None):
        self.verifiers = verifiers or []
    
    async def ensemble_verify(
        self,
        request: VerifyRequest,
        weights: List[float] = None
    ) -> Tuple[VerifyResponse, Dict[str, float]]:
        """
        集成多个验证器的结果
        
        Returns:
            (ensemble_result, individual_results)
        """
        if weights is None:
            weights = [1.0] * len(self.verifiers)
        
        # 并行验证
        tasks = [verifier.verify_draft(request) for verifier in self.verifiers]
        results = await asyncio.gather(*tasks)
        
        # 简单投票策略：选择接受率最高的
        best_result = max(results, key=lambda r: r.acceptance_rate)
        
        individual_rates = {
            f'verifier_{i}': r.acceptance_rate 
            for i, r in enumerate(results)
        }
        
        return best_result, individual_rates
