vllm_llama_inference_framework/requirements.txt
# 核心依赖
aiohttp>=3.8.0
aiohttp-cors>=0.7.0
numpy>=1.21.0
pyyaml>=6.0

# 可选依赖 (根据实际使用的引擎)
# vLLM (云端)
# vllm>=0.2.0

# llama.cpp Python 绑定 (边端)
# llama-cpp-python>=0.2.0

# 开发依赖 (可选)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0

# 性能监控 (可选)
psutil>=5.9.0
prometheus-client>=0.15.0
